Aesthetic and separation of concerns changes
