+++ apps/analytic_functions.py
@@ -6,10 +6,38 @@ import json
 import re
 import spacy
 
+from rich import print
+
 LEGAL_TESTS = "./data/legal-tests.json"
-with open(LEGAL_TESTS, "r") as file:
+with open(LEGAL_TESTS, "r", encoding="utf-8") as file:
     legal_tests = json.load(file)
 
+def local_text_analysis(citations: list):
+    """
+    Analyzes a text locally using the local analysis function. This function
+    returns a dictionary of the legal tests used in the text. If the function
+    finds a citation matching one linked to a legal test, it will return the
+    legal test.
+    """
+    print("\n[underline #FFA500]Analysis[/underline #FFA500]")
+    print("Analyzing text: ", end="")
+    legal_tests_identified = get_legal_test(citations)
+    print("[green]Done.[/green]")
+
+    text_string = ""
+
+    if legal_tests_identified:
+        print("\nTests found:")
+        text_string += "\nTests found:\n"
+
+        for test in legal_tests_identified:
+            print(f" \u2022 {test['short_form']}")
+            text_string += f" \u2022 {test['short_form']}\n"
+    else:
+        print("\nTests found: [red]None[/red]")
+        text_string += "\nTests found: None\n"
+
+    return text_string
 
 def retrieve_citations(text: str) -> dict:
     """
@@ -26,7 +54,9 @@ def retrieve_citations(text: str) -> dict:
         "Charter",
         "Charter of Rights and Freedoms",
         "Canadian Charter of Rights and Freedoms",
-        "Canadian Charter of Rights and Freedoms, Part I of the Constitution Act, 1982, being Schedule B to the Canada Act 1982 (UK), 1982, c 11",
+        "Canadian Charter of Rights and Freedoms, Part I of the Constitution \
+            Act, 1982, being Schedule B to the Canada Act 1982 (UK), 1982, c \
+            11",
         "Canadian Charter of Rights and Freedoms, Part I of the Constitution Act, 1982",
         "CDSA",
         "Controlled Drugs and Substances Act",
@@ -84,8 +114,29 @@ def get_legal_test(citations: dict) -> list[dict]:
     legal test.
     """
     case_citations = citations[0]["citations"]
-    
+
     for test in legal_tests:
         if test["origins"]["citation"] in case_citations:
             return [test]
 
+
+def detect_legal_tests(citations: dict):
+    """
+    Scans a list of citations to determine if any of them correspond to a legal
+    test.
+    """
+    print("\n[underline #FFA500]Analysis[/underline #FFA500]")
+    print("Analyzing text: ", end="")
+    legal_tests_detected = get_legal_test(citations)
+    print("[green]Done.[/green]")
+    test_list = []
+
+    if legal_tests_detected:
+        print("\nTests found:")
+        for test in legal_tests_detected:
+            print(f" \u2022 {test['short_form']}")
+            test_list.append(test["short_form"])
+    else:
+        print("\nTests found: [red]None[/red]")
+
+    return test_list
