+++ firac.py
@@ -14,14 +14,14 @@ import json
 
 from typing import Optional
 from rich import print
-from prettytable import PrettyTable
 
 import typer
 
-from apps.analytic_functions import retrieve_citations, get_legal_test
+from apps.analytic_functions import detect_legal_tests, local_text_analysis
 from apps.classification_functions import classify_firac
-from apps.html_to_txt import canlii_html_to_txt, write_text
-from apps.summarization_functions import extraction_text_summarizer, preprocess_text_for_gpt
+from apps.extraction_functions import extract_text, extract_citations, verify_file_path
+from apps.html_to_txt import write_text
+from apps.summarization_functions import preprocess_text_for_gpt, local_text_summary
 from apps.gpt_functions import gpt_hybrid_analysis_manual
 
 
@@ -119,18 +119,18 @@ def create_firac(file: Optional[str] = None):
     if file is None:
         # Prompt the user to select a file, input text manually, or exit.
         user_input = input("(S)elect a file, (I)nteractive mode, or E(x)it?\n")
-        
+
         if user_input.lower() == "s":
             file = input("Enter the file path: ")
             create_firac(file)
-        
+
         elif user_input.lower() == "i":
             summary = firac_manual()
             analysis = local_text_analysis(summary)
             report = gpt_hybrid_analysis_manual(summary)
-            
+
             return report, analysis
-        
+
         elif user_input.lower() == "x":
             sys.exit()
     # Check the file suffix to determine what functions to run.
@@ -144,7 +144,7 @@ def create_firac(file: Optional[str] = None):
             report = gpt_hybrid_analysis_manual(summary)
 
             return report, analysis
-        
+
         elif file.endswith(".json"):
             # Load the summary file
             with open(file, "r", encoding="utf-8") as file:
@@ -159,106 +159,13 @@ def create_firac(file: Optional[str] = None):
             report = gpt_hybrid_analysis_manual(summary)
 
             return report, analysis
-        
+
         else:
             print("[bold red]File type not supported.[/bold red]")
             create_firac()
 
 
 # Functions
-# Text extraction and file verification
-
-def verify_file_path(file_path: str):
-    """
-    Verifies that a file exists at the specified path.
-    """
-    print("Verifying file path: ", end="")
-    if not os.path.exists(file_path):
-        print("File not found.")
-        sys.exit()
-    else:
-        print("[green]Done.[/green]")
-
-def extract_text(file_path: str):
-    """
-    Extracts text from an HTML file.
-    """
-    verify_file_path(file_path)
-
-    if file_path.endswith(".html"):
-        text = canlii_html_to_txt(file_path)
-        write_text(text, file_path)
-        return text
-
-    elif file_path.endswith(".txt"):
-        with open(file_path, "r", encoding="utf-8") as file:
-            text = file.read()
-        print("[green]Text file found:[/green] returning file contents.")
-        return text
-
-    else:
-        print("[bold red]File type not supported.[/bold red]")
-        sys.exit()
-
-
-def extract_citations(text: str):
-    """
-    Extracts citations from a legal text, if any, and exports the results as a
-    dictionary.
-    """
-    print("\n[underline #FFA500]Citation extraction[/underline #FFA500]")
-    print("Retrieving citations: ", end="")
-    citations = retrieve_citations(text)
-    print("[green]Done.[/green]\n")
-
-    # Add citations to a list if they contain "SCC"
-    scc_citations = [
-        citation for citation in citations[0]["citations"] if "SCC" in citation
-    ]
-    # Add citations to another list if they contain "CA"
-    ca_citations = [
-        citation for citation in citations[0]["citations"] if "CA" in citation
-    ]
-    unknown_citations = [
-        citation for citation in citations[0]["citations"] if "CanLII" in citation
-    ]
-    other_citations = [
-        citation
-        for citation in citations[0]["citations"]
-        if "SCC" not in citation and "CA" not in citation and "CanLII" not in citation
-    ]
-
-    print(f"\u2022 Total case citations found: {len(citations[0]['citations'])}")
-    print(f"\u2022 Supreme Court of Canada citations: {len(scc_citations)}")
-    print(f"\u2022 Court of Appeal citations: {len(ca_citations)}")
-    print(f"\u2022 Provincial and superior court decisions: {len(other_citations)}")
-    print(f"\u2022 CanLII citations (no jurisdiction/court listed): {len(unknown_citations)}")
-
-    return citations
-
-
-def detect_legal_tests(citations: dict):
-    """
-    Scans a list of citations to determine if any of them correspond to a legal
-    test.
-    """
-    print("\n[underline #FFA500]Analysis[/underline #FFA500]")
-    print("Analyzing text: ", end="")
-    legal_tests = get_legal_test(citations)
-    print("[green]Done.[/green]")
-    test_list = []
-
-    if legal_tests:
-        print("\nTests found:")
-        for test in legal_tests:
-            print(f" \u2022 {test['short_form']}")
-            test_list.append(test["short_form"])
-    else:
-        print("\nTests found: [red]None[/red]")
-
-    return test_list
-
-
 def firac_manual():
     """
     Takes user input to create the FIRAC structure. Returns the structure as a
@@ -291,7 +198,8 @@ def firac_manual():
             conclusion = input("Enter conclusion: ")
             firac["opinions"]["conclusion"] = conclusion
         elif user_input.lower() == "q":
-            opinion = input("Enter the opinion type ([M]ajority; [D]issenting; [Con]curring; [Cur]iam; [Cou]rt): ")
+            opinion = input("Enter the opinion type ([M]ajority; [D]issenting;\
+                             [Con]curring; [Cur]iam; [Cou]rt): ")
             # Validate the opinion type
             if opinion.lower() == "m":
                 firac["opinions"]["opinion_type"] = "majority"
@@ -337,77 +245,6 @@ def firac_json():
         firac = json.load(file)
 
     return firac
-
-
-def local_text_summary(firac: dict) -> dict:
-    """
-    Summarizes a text locally using the local summarization function. This
-    function ranks sentences based on a simple word frequency algorithm. Future
-    verions will allow more sophisticated summarization methods.
-    """
-    print("\n[bold underline #FFA500]Summarization[/bold underline #FFA500]")
-    print("Summarizing text: ", end="")
-    summary = {}
-    table = PrettyTable()
-    table.field_names = ["", "Total spaCy Tokens", "Percentage Included"]
-
-    def process_key(key: str):
-        """
-        Processes the key and returns the summary.
-        """
-        # Remove extraneous spaces and characters
-        firac[key] = preprocess_text_for_gpt(firac[key])
-
-        # Summarize the text
-        firac[key] = extraction_text_summarizer(firac[key])
-
-        # Add the summary to the summary dictionary
-        summary[key] = firac[key]
-        table.add_row(
-            [f"{key.title()}", round(firac[key][1] * 100, 2), firac[key][2]]
-        )
-
-    # Each FIRAC key contains a list of sentences. Go through each list and
-    # remove \n characters. Then, join the sentences into a single string.
-    
-    keys = ["opinion_type", "facts", "issues", "rules", "analysis", "conclusion"]
-
-    for key in keys:
-        process_key(key)
-
-
-    print("[green]Done.[/green]\n")
-
-    typer.echo(table)
-    return summary
-
-
-def local_text_analysis(citations: list):
-    """
-    Analyzes a text locally using the local analysis function. This function
-    returns a dictionary of the legal tests used in the text. If the function
-    finds a citation matching one linked to a legal test, it will return the
-    legal test.
-    """
-    print("\n[underline #FFA500]Analysis[/underline #FFA500]")
-    print("Analyzing text: ", end="")
-    legal_tests = get_legal_test(citations)
-    print("[green]Done.[/green]")
-
-    text_string = ""
-
-    if legal_tests:
-        print("\nTests found:")
-        text_string += "\nTests found:\n"
-
-        for test in legal_tests:
-            print(f" \u2022 {test['short_form']}")
-            text_string += f" \u2022 {test['short_form']}\n"
-    else:
-        print("\nTests found: [red]None[/red]")
-        text_string += "\nTests found: None\n"
-
-    return text_string
 
 
 # Calls the app
